third_party/blink/web_tests/external/wpt/html/canvas/offscreen/text/2d.text.drawing.style.textRendering.settings-expected.txt

This is a testharness.js-based test.
FAIL Testing basic functionalities of textRendering in Canvas assert_equals: ctx.textRendering === "auto" (got optimizeSpeed[string], expected auto[string]) expected "auto" but got "optimizeSpeed"
Harness: the test ran to completion.
